Update the chain_spec to use a separate file for paseo-local
